(prf): send bytes directly via client
