+++ src/core/mod.rs
@@ -1,3 +1,4 @@
 pub mod derived;
 pub mod mail;
+pub mod py_interface;
 pub mod response;
